Move utility to AppStatusStoreUtils
