+++ core/src/main/scala/org/apache/spark/util/Utils.scala
@@ -199,18 +199,6 @@ private[spark] object Utils extends Logging with SparkClassUtils {
     Try { classForName(clazz, initialize = false) }.isSuccess
   }
 
-  def getQuantilesValue(
-    values: IndexedSeq[Double],
-    quantiles: Array[Double]): IndexedSeq[Double] = {
-    val count = values.size
-    if (count > 0) {
-      val indices = quantiles.map { q => math.min((q * count).toLong, count - 1) }
-      indices.map(i => values(i.toInt)).toIndexedSeq
-    } else {
-      IndexedSeq.fill(quantiles.length)(0.0)
-    }
-  }
-
   /**
    * Run a segment of code using a different context class loader in the current thread
    */
